Fix sign/verify when protected header is absent.
Previously, if there was no protected header, we would attempt to serialize a
nil pointer. json.Marshal would happily serialize it as the literal string
"null", which would cause an incorrect signature or a verify error.

Instead, if there is no protected header, explicitly set the protected part of
the input to be signed to the empty string.