Optionally disable embedding JWK in protected header
In some applications, the signer's public key is already known to
the recipient, or the 'kid' property in the header can be used to
lookup the key. Simply speaking, embedding the signer's public key
in the protected header's 'jwk' property is sometimes not wanted.

This commit adds a new method to the Signing interface to disable
embedding of the public key of the signer. It will be embedded by
default to maintain backwards compatibility.