Preserve integers when normalizing JWT claims (fixes #142)
We normalize claims for tokens given to us in the builder by marshaling
and then unmarshaling as a JSON object. However, because JSON officially
only supports floating-point numbers, this loses precision for integers
embedded in structs. We switch to using json.Decoder (with UseNumber
option) to decode numbers into json.Number, which preserves the precise
value for both integers and floats.