Expose payload verifier in interface
Makes the previously private payloadVerifier interface public as just
a Verifier, and merges it with OpaqueVerifier. Adds type-safe functions
for instantiating new verifiers from RSA, ECDSA, etc. keys so we can
avoid some of the reflection we used to do internally. Should also make
the interface friendlier to consumers of the library.